+++ platform/swift/source/Logger.swift
@@ -37,6 +37,9 @@ public final class Logger {
     private(set) var dispatchSourceMemoryMonitor: DispatchSourceMemoryMonitor?
     private(set) var resourceUtilizationTarget: ResourceUtilizationController
     private(set) var eventsListenerTarget: EventSubscriber
+    private let appStateAttributes: AppStateAttributes
+    private let localeAttributes: LocaleAttributes
+    private let networkAttributes: NetworkAttributes
 
     private let sessionURLBase: URL
     private var crashReporterService: CrashReporterService?
@@ -136,10 +139,11 @@ public final class Logger {
         self.timeProvider = timeProvider
         let start = timeProvider.uptime()
 
-        let appStateAttributes = AppStateAttributes()
+        self.appStateAttributes = AppStateAttributes()
         let clientAttributes = ClientAttributes()
         let deviceAttributes = DeviceAttributes()
-        let networkAttributes = NetworkAttributes()
+        self.localeAttributes = LocaleAttributes()
+        self.networkAttributes = NetworkAttributes()
 
         let metadataProvider = MetadataProviderController(
             dateProvider: dateProvider ?? SystemDateProvider(),
@@ -178,9 +182,9 @@ public final class Logger {
             bufferDirectoryPath: directoryURL.path,
             sessionStrategy: sessionStrategy,
             metadataProvider: metadataProvider,
-            initialOotbFields: appStateAttributes.initialOotbFields()
-                + deviceAttributes.initialOotbFields()
-                + networkAttributes.initialOotbFields(),
+            initialOotbFields: self.appStateAttributes.initialOotbFields()
+                + self.localeAttributes.initialOotbFields()
+                + self.networkAttributes.initialOotbFields(),
             // TODO(Augustyniak): Pass `resourceUtilizationTarget`, `sessionReplayTarget`,
             // and `eventsListenerTarget` as part of the `self.underlyingLogger.start()` method call instead.
             // Pass the event listener target here and finish setting up
@@ -222,7 +226,7 @@ public final class Logger {
 
         self.eventsListenerTarget.setUp(
             logger: self.underlyingLogger,
-            appStateAttributes: appStateAttributes,
+            appStateAttributes: self.appStateAttributes,
             clientAttributes: clientAttributes,
             timeProvider: timeProvider
         )
@@ -237,8 +241,9 @@ public final class Logger {
         // Start attributes before the underlying logger is running to increase the chances
         // of out-of-the-box attributes being ready by the time logs emitted as a result of the logger start
         // are emitted.
-        deviceAttributes.start(with: self.underlyingLogger)
-        networkAttributes.start(with: self.underlyingLogger)
+        self.appStateAttributes.start(with: self.underlyingLogger)
+        self.localeAttributes.start(with: self.underlyingLogger)
+        self.networkAttributes.start(with: self.underlyingLogger)
 
         self.underlyingLogger.start()