+++ test/platform/swift/unit_integration/core/LocaleAttributesTests.swift
@@ -10,20 +10,17 @@ import CaptureMocks
 import Foundation
 import XCTest
 
-final class DeviceAttributesTests: XCTestCase {
+final class LocaleAttributesTests: XCTestCase {
     func testLocale() {
-        let deviceAttributes = DeviceAttributes()
-        // Confirm that attributes are initialized with the `locale` field value.
-        let localeStr = deviceAttributes.initialOotbFields()[0].data as! String
+        let localeAttributes = LocaleAttributes()
+        let localeStr = localeAttributes.initialOotbFields()[0].data as! String
         XCTAssertNotNil(Locale(identifier: localeStr))
 
         let logger = MockCoreLogging()
-        deviceAttributes.start(with: logger)
-        // Confirm that the `locale` field looks OK after starting device attributes.
+        localeAttributes.start(with: logger)
+
         let nowLocaleStr = logger.ootbFields["_locale"]!
         XCTAssertNotNil(Locale(identifier: nowLocaleStr))
-
-        // And is the same as before.
         XCTAssertEqual(localeStr, nowLocaleStr)
     }
 }